backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.7",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "djoser>=2.3.3",
    "drf-yasg>=1.21.11",
    "openpyxl>=3.1.5",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
    "python-decouple>=3.8",
    "opencv-python>=4.8.1",
    "deepface>=0.0.93",
    "tf-keras>=2.18.0",
    "numpy>=1.26.0",
    "pgvector>=0.4.1",
    "scikit-learn>=1.3.0",
]
